+++ CHANGES.md
@@ -2,6 +2,8 @@
 
 #### Added
 
+- Allow to explicitly set the kind of blocks in labels: `ocaml`, `cram`, `toplevel` or `include`. (#237, @gpetiot)
+
 #### Changed
 
 #### Deprecated